[lldb/test] Split heavy swift API tests into per-file parts
Several swift API test classes generate enough variants
(debug_info × swift_module_importer × swift_embedded) that the
per-test wall-clock blows past lit's 600s file-level timeout on
remote-iOS, even when the work each individual variant does is fast.
Split each affected class into multiple part files so dotest sees each
part as its own test entry, and the timeout budget gets divided
across them.

Layout: parts are written as siblings of the original test file in
the same directory. Each <Stem>_partNN.py keeps the original class
name, all decorators, all non-test members (setUp, tearDown, helpers),
trailing in-method comments (FileCheck CHECK lines), and any
module-level helpers defined after the class. The aux files
(Makefile, main.swift / main.cpp / *.h / module.modulemap / libs/ ...)
are already in the parent dir, so no symlinks or subdirs are needed.

This avoids the Windows MAX_PATH (260) issue that an earlier subdir
layout hit on the deepest paths.

The parent Test*.py files are removed in the same change so dotest
only discovers the per-part files and each part gets its own 600s lit
timeout budget.

Splits applied (parts × methods):

  cxx_interop/backward/stepping
    TestSwiftBackwardInteropSteppingClass        3 × 4 = 12
    TestSwiftBackwardInteropSteppingStruct       3 × {4,4,2} = 10
  cxx_interop/forward/stepping-forward-interop
    TestSwiftForwardInteropStepping              3 × {4,4,2} = 10
  cxx_interop/forward/stepping-forward-interop_header_only
    TestSwiftForwardInteropSteppingHeaderOnly    3 × {4,4,2} = 10
  explicit_modules/bridgingheader_caching
    TestSwiftExplicitModulesBridgingHeaderCaching 2 × 1 = 2
  explicit_modules/simple
    TestSwiftExplicitModules                     2 × {2,1} = 3
  parseable_interfaces/shared
    TestSwiftInterfaceNoDebugInfo                3 × 1 = 3
  parseable_interfaces/static
    TestSwiftInterfaceStaticNoDebugInfo          2 × 1 = 2
  po/pointer_and_mangled_typename
    TestSwiftPrintObjectPointerAndTypeName       3 × {4,4,3} = 11
